+++ examples/table_movie.py
@@ -111,7 +111,7 @@ with Live(table_centered, console=console, screen=False, refresh_per_second=20):
     with beat(10):
         table.show_footer = True
 
-    table_width = Measurement.get(console, table, console.width).maximum
+    table_width = console.measure(table).maximum
 
     with beat(10):
         table.columns[2].justify = "right"
@@ -175,7 +175,7 @@ with Live(table_centered, console=console, screen=False, refresh_per_second=20):
     with beat(10):
         table.pad_edge = False
 
-    original_width = Measurement.get(console, table).maximum
+    original_width = console.measure(table).maximum
 
     for width in range(original_width, console.width, 2):
         with beat(1):