+++ rich/spinner.py
@@ -1,30 +1,30 @@
 from typing import cast, List, Optional, TYPE_CHECKING
 
 from ._spinners import SPINNERS
-from .console import Console
 from .measure import Measurement
-from .style import StyleType
-from .text import Text, TextType
+from .table import Table
+from .text import Text
 
 if TYPE_CHECKING:
-    from .console import Console, ConsoleOptions, RenderResult
+    from .console import Console, ConsoleOptions, RenderResult, RenderableType
+    from .style import StyleType
 
 
 class Spinner:
     def __init__(
         self,
         name: str,
-        text: TextType = "",
+        text: "RenderableType" = "",
         *,
-        style: Optional[StyleType] = None,
-        speed: float = 1.0,
+        style: Optional["StyleType"] = None,
+        speed=1.0,
     ) -> None:
         """A spinner animation.
 
         Args:
             name (str): Name of spinner (run python -m rich.spinner).
-            text (TextType, optional): Text to display at the right of the spinner. Defaults to "".
-            style (StyleType, optional): Style for sinner amimation. Defaults to None.
+            text (RenderableType, optional): A renderable to display at the right of the spinner (str or Text typically). Defaults to "".
+            style (StyleType, optional): Style for spinner animation. Defaults to None.
             speed (float, optional): Speed factor for animation. Defaults to 1.0.
 
         Raises:
@@ -40,16 +40,13 @@ class Spinner:
         self.start_time: Optional[float] = None
         self.style = style
         self.speed = speed
-        self.time = 0.0
+        self.frame_no_offset: float = 0.0
+        self._update_speed = 0.0
 
     def __rich_console__(
         self, console: "Console", options: "ConsoleOptions"
     ) -> "RenderResult":
-        time = console.get_time()
-        if self.start_time is None:
-            self.start_time = time
-        text = self.render(time - self.start_time)
-        yield text
+        yield self.render(console.get_time())
 
     def __rich_measure__(
         self, console: "Console", options: "ConsoleOptions"
@@ -57,18 +54,60 @@ class Spinner:
         text = self.render(0)
         return Measurement.get(console, options, text)
 
-    def render(self, time: float) -> Text:
+    def render(self, time: float) -> "RenderableType":
         """Render the spinner for a given time.
 
         Args:
             time (float): Time in seconds.
 
         Returns:
-            Text: A Text instance containing animation frame.
+            RenderableType: A renderable containing animation frame.
         """
-        frame_no = int((time * self.speed) / (self.interval / 1000.0))
-        frame = Text(self.frames[frame_no % len(self.frames)], style=self.style or "")
-        return Text.assemble(frame, " ", self.text) if self.text else frame
+        if self.start_time is None:
+            self.start_time = time
+
+        frame_no = ((time - self.start_time) * self.speed) / (
+            self.interval / 1000.0
+        ) + self.frame_no_offset
+        frame = Text(
+            self.frames[int(frame_no) % len(self.frames)], style=self.style or ""
+        )
+
+        if self._update_speed:
+            self.frame_no_offset = frame_no
+            self.start_time = time
+            self.speed = self._update_speed
+            self._update_speed = 0.0
+
+        if not self.text:
+            return frame
+        elif isinstance(self.text, (str, Text)):
+            return Text.assemble(frame, " ", self.text)
+        else:
+            table = Table.grid(padding=1)
+            table.add_row(frame, self.text)
+            return table
+
+    def update(
+        self,
+        *,
+        text: "RenderableType" = "",
+        style: Optional["StyleType"] = None,
+        speed: Optional[float] = None,
+    ) -> None:
+        """Updates attributes of a spinner after it has been started.
+
+        Args:
+            text (RenderableType, optional): A renderable to display at the right of the spinner (str or Text typically). Defaults to "".
+            style (StyleType, optional): Style for spinner animation. Defaults to None.
+            speed (float, optional): Speed factor for animation. Defaults to None.
+        """
+        if text:
+            self.text = text
+        if style:
+            self.style = style
+        if speed:
+            self._update_speed = speed
 
 
 if __name__ == "__main__":  # pragma: no cover