+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed initial blank lines removed from Syntax https://github.com/willmcgugan/rich/issues/1214
 - Added Console.measure as a convenient alias for Measurement.get
 
+### Fixed
+
+- Colors are now shown in the Spyder console
+
 ## [10.1.0] - 2020-04-03
 
 ### Fixed